% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBA_functions.R
\name{detrend}
\alias{detrend}
\title{Linearly Detrend a Data Series}
\usage{
detrend(vec, std)
}
\arguments{
\item{vec}{Vector of ...}

\item{std}{TRUE/FALSE Boolean: Should the variance of each stationary block be set to one across all blocks?}
}
\value{
vector of of the detrended vec parameter
}
\description{
Linearly Detrend a Data Series
}
