% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBA_functions.R
\name{eba.b}
\alias{eba.b}
\title{Finds the frequency partition point using Hochberg step up rule}
\usage{
eba.b(X.dm, f, startf, endf, covg, alpha)
}
\arguments{
\item{alpha}{}
}
\description{
Finds the frequency partition point using Hochberg step up rule
}
