% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBA_functions.R
\name{eba.flat}
\alias{eba.flat}
\title{Tests for a flat spectrum through time (i.e. frequency component contribution same across time)}
\usage{
eba.flat(f, partfinal, ghat, covg)
}
\arguments{
\item{f}{Fourier frequencies}

\item{partfinal}{final partition using EBA algorithm}

\item{ghat}{multitaper spectral estimates (N x B)}

\item{covg}{estimated covariance of demeaned multitaper spectral estimates for b=1,...B (N x N x B)}
}
\value{
P-Value for test
}
\description{
Tests for a flat spectrum through time (i.e. frequency component contribution same across time)
}
