% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fhat_lp}
\alias{fhat_lp}
\title{ADD TITLE}
\usage{
fhat_lp(X, N, stdz)
}
\arguments{
\item{X}{ADD DESCRIPTION}

\item{N}{ADD DESCRIPTION}

\item{stdz}{ADD DESCRIPTION}
}
\value{
ADD DESCRIPTION
}
\description{
ADD DESCRIPTION
}
\examples{
# ADD EXAMPLE
}
