% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runApp.R
\name{launchApp}
\alias{launchApp}
\title{Launch a Shiny Application that houses all of the algorithms seen in this package}
\usage{
launchApp()
}
\value{
Running this function will launch this shiny application in a new window.
}
\description{
This application serves to provide a way to find and analyze frequency bands in Univariate, Multivariate, and Functional data, in a graphic-based, user-friendly manner.
}
\details{
There is a vignette entitled "ShinyAppUsage" that is included with this package, which goes into much greater detail about how to use and understand the application.
}
\examples{
launchApp()
}
\references{
(Univariate): doi.org/10.1080/01621459.2019.1671199 \cr \cr
(Multivariate): doi.org/10.48550/arXiv.2301.03664 \cr \cr
(Functional): doi.org/10.48550/arXiv.2102.01784 \cr \cr
}
