% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mEBA_Rfunctions.R
\name{meba.simdata}
\alias{meba.simdata}
\title{function to simulate data for 5 different settings}
\usage{
meba.simdata(t)
}
\arguments{
\item{t}{ADD DESCRIPTION}
}
\value{
A named list containing:
\describe{
  \item{wn}{DESCRIPTION}
  \item{bL}{DESCRIPTION}
  \item{bS}{DESCRIPTION}
  \item{bL2f15}{DESCRIPTION}
  \item{bL2f35}{DESCRIPTION}
  \item{bS2f15}{DESCRIPTION}
  \item{bS2f35}{DESCRIPTION}
}
}
\description{
function to simulate data for 5 different settings
}
\examples{
set.seed(836583)
sim <- meba.simdata(400)
names(sim)
sapply(sim, length)
}
