% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{msboot}
\alias{msboot}
\title{ADD TITLE}
\usage{
msboot(nrep, x, Wsel, stdz, ncore)
}
\arguments{
\item{nrep}{ADD DESCRIPTION}

\item{x}{ADD DESCRIPTION}

\item{Wsel}{ADD DESCRIPTION}

\item{stdz}{ADD DESCRIPTION}

\item{ncore}{ADD DESCRIPTION}
}
\value{
ADD DESCRIPTION
}
\description{
ADD DESCRIPTION
}
\examples{
# ADD EXAMPLE
}
