% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBA_functions.R
\name{eba.search}
\alias{eba.search}
\title{Title}
\usage{
eba.search(X, N, K, std, alpha)
}
\arguments{
\item{X}{numeric vector with a length > 0 and not missing or non-finite values}

\item{N}{number of of observations per approximately stationary block}

\item{K}{number of tapers to use in multitaper spectral estimator}

\item{std}{TRUE/FALSE Boolean: should the variance of each stationary block be set to one across all blocks?}

\item{alpha}{significance level to use for testing partition points using FRESH statistic}
}
\value{
Results on partition of frequency space
}
\description{
Title
}
