% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EBA_functions.R
\name{eba.simdata}
\alias{eba.simdata}
\title{Function to simulate time series data with length 'T'}
\usage{
eba.simdata(T)
}
\arguments{
\item{T}{total length of intended time series}
}
\value{
List of 3 frequency series; wn (white noise), bl, and bs (time series processes in the data, 3 different settings in the paper)
}
\description{
Function to simulate time series data with length 'T'
}
\examples{
eba.simdata(T=50000)
}
