% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EFBA-package.R
\docType{package}
\name{EFBA-package}
\alias{EFBA}
\alias{EFBA-package}
\title{EFBA: Empirical Frequency Band Analysis}
\description{
Tools for empirical frequency-band analysis of non-stationary time series.
These imports make base plotting and IO functions visible to the package,
removing "no visible global function definition" NOTES.
}
\author{
\strong{Maintainer}: Matt Byrom \email{matthew.james.byrom@tamu.edu}

Authors:
\itemize{
  \item Scott Bruce \email{sabruce@tamu.edu}
}

Other contributors:
\itemize{
  \item Dylan Ward [contributor]
  \item Mohit Chhaparia \email{mohit_chhaparia@tamu.edu} [contributor]
}

}
\keyword{internal}
