% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fEBA_Rfns.R
\name{fws.sim}
\alias{fws.sim}
\title{Simulate Functional White Noise Data}
\usage{
fws.sim(nb = 15, gsz = 20, Ts, seed)
}
\arguments{
\item{nb}{Number of basis functions to use when simulating the data. Default value is 15}

\item{gsz}{Number of points along the functional domain at which to evaluate these basis functions. Default value is 20}

\item{Ts}{Numeric/integer with total length of intended time series}

\item{seed}{The seed you wish to set, to ensure reproducibility}
}
\value{
A Ts x gsz dimension matrix, containing the simulated values
}
\description{
The function fws.sim will simulate a functional time series of white noise data
}
\details{
If no values for nb or gsz are selected, the function will utilize the aforementioned default values of 15 and 20, respectively.
Ts and seed need to be provided every time the function is run.
Every input must be a numeric, positive number.
}
\examples{
fws.sim(Ts = 2000, seed = 234)
}
