% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mEBA_Rfunctions.R
\name{preprocess}
\alias{preprocess}
\title{function to detrend, standardize, and filter out low frequencies (below 1 Hz)}
\usage{
preprocess(sgnl, dsfrq, channels)
}
\arguments{
\item{sgnl}{ADD DESCRIPTION}

\item{dsfrq}{ADD DESCRIPTION}

\item{channels}{ADD DESCRIPTION}
}
\value{
ADD DESCRIPTION
}
\description{
function to detrend, standardize, and filter out low frequencies (below 1 Hz)
}
\examples{
# ADD EXAMPLE
}
